Implémentation des nouvelles fonctionnalités

+++ src/Controller/Admin/MediaController.php
@@ -12,6 +12,7 @@ use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\HttpFoundation\File\UploadedFile;
 use Symfony\Bridge\Doctrine\Attribute\MapEntity;
+use Symfony\Component\Form\FormError;
 
 class MediaController extends AbstractController
 {
@@ -33,7 +34,7 @@ class MediaController extends AbstractController
             25 * ($page - 1)
         );
 
-        $total = count($mediaRepository->findAll());
+        $total = count($medias);
 
         return $this->render('admin/media/index.html.twig', [
             'medias' => $medias,
@@ -46,24 +47,43 @@ class MediaController extends AbstractController
     public function add(Request $request, ManagerRegistry $doctrine): Response
     {
         $media = new Media();
-        $form = $this->createForm(MediaType::class, $media, ['is_admin' => $this->isGranted('ROLE_ADMIN')]);
+        return $this->handleMediaForm($request, $doctrine, $media);
+    }
+
+    #[Route('/admin/media/update/{id}', name: 'admin_media_update')]
+    public function update(Request $request, #[MapEntity] Media $media, ManagerRegistry $doctrine): Response
+    {
+        $this->denyAccessUnlessGranted('EDIT', $media);
+        return $this->handleMediaForm($request, $doctrine, $media);
+    }
+
+
+    private function handleMediaForm(Request $request, ManagerRegistry $doctrine, Media $media): Response
+    {
+        $originalPath = $media->getPath();
+
+        $form = $this->createForm(MediaType::class, $media, [
+            'is_admin' => $this->isGranted('ROLE_ADMIN'),
+        ]);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
-            if (!$this->isGranted('ROLE_ADMIN')) {
-                $user = $this->getUser();
-                if ($user instanceof User) {
-                    $media->setUser($user);
-                }
-            }
-
-
             /** @var UploadedFile|null $file */
-            $file = $media->getFile();
+            $file = $form->get('file')->getData();
+
             if ($file !== null) {
-                $filename = 'uploads/' . md5(uniqid((string) random_int(0, 1000), true)) . '.' . $file->guessExtension();
+                if ($originalPath && file_exists($originalPath)) {
+                    @unlink($originalPath);
+                }
+
+                $extension = $file->guessExtension() ?: $file->getClientOriginalExtension();
+                $filename = 'uploads/' . md5(uniqid((string) random_int(0, 1000), true)) . '.' . $extension;
                 $media->setPath($filename);
                 $file->move('uploads/', basename($filename));
+            }
+
+            if (!$this->isGranted('ROLE_ADMIN') && !$media->getUser()) {
+                $media->setUser($this->getUser());
             }
 
             $em = $doctrine->getManager();
@@ -75,12 +95,15 @@ class MediaController extends AbstractController
 
         return $this->render('admin/media/add.html.twig', [
             'form' => $form->createView(),
+            'media' => $media,
         ]);
     }
 
     #[Route('/admin/media/delete/{id}', name: 'admin_media_delete')]
     public function delete(#[MapEntity] Media $media, ManagerRegistry $doctrine): Response
     {
+        $this->denyAccessUnlessGranted("DELETE", $media);
+
         $path = $media->getPath();
         $em = $doctrine->getManager();
         $em->remove($media);
